Rework type class structure for distributions statistics
Some distributions do not have variance or mean for some values
of parameters. Student-T is an example. Simplest solution is
to return NaN but it's way too implict

One option is change Mean (and Variance) type class:

> class Mean d where
>   mean :: d → Maybe Double

But it would be really annoying to unwrap Just constructor
for distribution which have mean defined for all parameters.
It also encourages use of fromJust which is bad practice

So following type classes is used:

> class MaybeMean d where
>   maybeMean :: d → Maybe Double
>
> class MaybeMean d ⇒ Mean d where
>   mean :: d → Double
>   mean = fromJust . maybeMean

Obviously it would be better to implement maybeMean in terms of
mean since usually mean is always defined. Alas it's not possible.