Add Laplace Distribution.
The Laplace Distribution (also known as double-exponential distribution) arises
from Laplace's first error law (the distribution of the absolute deviations
from the mean) and its relationship with the Gaussian Distribution (normal
distribution) is the same as the relationship between L1 curve fitting (least
absolute deviations) and L2 curve fitting (least squares).

The distribution is widely used in differential privacy, being the core of the
Laplace Mechanism for sensitizing a dataset/query result. It is also used in
speech recognition.