Add laplaceFromSample method.
Allows estimating the parameters of a Laplace distribution based on
a sample (location is the sample median and scale is the average
mean deviation from the location). No checks are made to test that
the sample comes from a Laplace distribution.

With this, the Statistics.Distribution.Laplace module closely mimics the
Statistics.Distribution.Exponential one.