Added a two sample student t test along with documentation.
I tried to follow the format of the hypothesis test modules which
already exist in the statistics package (especially the ChiSquared
test). I grepped around the test tree, but did not encounter
any (software) tests for these sort of statistical hypothesis tests.
Without a template, I opted not to write any automated tests, and
instead simply sanity checked my implimentation against the
R builtin function t.test. If I was in error in thinking that
there are not currently any automated tests for the hypothesis
tests please reject this PR as I would like to supply one.

In light of issue #25 I provide both a fixed hypothesis test
function which can be specialized with a given alpha value,
and a more general function which simply computes the
p-value corresponding to the documented alternative hypothesis.

I focused mainly on correctness, but if there is anything
obvious I can do to improve performance I would love feedback.

I also added a cookie cutter haskell .gitignore.