Regression functions added to address point 1. in issue #67
Added two type classes, 'WeightedNormalRegress' and 'NormalRegress',
which have primary methods 'weightedNormalRegress' and 'normalRegress',
respectively, which run weigihted and ordinary least squares, again
respectively. These methods extend 'olsRegress' and compute regression
coefficient standard errors and an overall model fit test statistic
in addition to estimating the coefficients and returning the R-squared
goodness of fit metric. Both type classes have instances for when
the noise variance is unknown and when it is known. In both instances,
it is assumed that the noise is Gaussian.

Some auxiliary changes were needed to make these extensions to the
Regression module:

**(list below is of important addtions but is not exhaustive)**

  - A new error data type, 'TErr', which stores t-distributed errors
    was added to Statistics/Types.hs for coefficient standard errors
    when the noise variance is unknown.

  - A function 'diagOf' in the Statistics/Matrix.hs returns the
    diagonal of a square matrix. This is used to retrieve the standard
    errors of each regression coefficient in 'varCoeff'.

  - 'inv' in Statistics/Regression.hs can take the inverse of upper
    or lower triangular matrices (also needed in 'varCoeff'.

Some future-looking changes were made as well:

  - In anticipation of implementing generalized least square which
    allows for specifcation of arbitrary linear correlation structure
    across the noise component, this commit has implemented the
    Cholesky factorization 'chol' in Statistics/Matrix/Algorithms.hs.

  - 'solve' can now handle lower triangular matrices as well.