Fix for loop error in solve method for lower triangular matrices.
Previously extension of `solve` in Regression module to deal with
lower triangular matrices incorrectly used `rfor` for a forward
for loop. To resolve this issue and to still allow `solve` to work
for both with maximum code reusability I wrote a wrapper for `rfor`
and `for` in the Functions modules called `for_` that checks the
the start and end indices to determine which direction to run the
loop.