Exponential Distribution:  Fix fromSample, take 2.
There were a number of errors in the fromSample function for exponential
distributions.

First, the formula was wrong:  the exponential distribution is here
parametrized by the rate, and not by the mean.  So, e.g. if you fed it
samples from an (ED 2.0) it would return you something close to an (ED
0.5) instead.

Second, it rejected any sample that had negative values in it.  This is
not helpful: While it's correct that exact exponential variates cannot
be negative, real-world samples can be noisy and may contain
negative data points.