FromSample:  Revise documentation.
Revise the documentation of the type class and some instances to better
reflect the intent that a fitted distribution should be returned
whenever it is reasonably possible to do so.

For Laplace, remove the comment about checking for the distribution
since it may be misleading.  Instead summarize what quantities are used
for the estimate.