Geometric distribution:  Improve accuracy for large s.
When the success probability parameter s >= 0.5, the approach using
log1p and similar functions is not as accurate as the previously used
direct computation.  This is because in that case, the intermediate
quantity 1 - s can be computed without rounding error.

So we now use the direct computation for s >= 0.5 and the log1p approach
for s < 0.5.

With this change, the relative error seems to be < 1e-14 for s >= 0.5.